change from CTest to custom test target
CTest cannot build the test command before running it without any of
the well known nasty workarounds, the best resulting in a new target
doing the build and calling the test target afterwards. Since all we
have to do is running the test binary, we can just have a custom test
target and ignore CTest.